move ndk init from app to layout and add routes for relay and group
